requirements.txt
alembic==1.8.1
anyio==3.6.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.0.1
fastapi==0.86.0
fastapi-utils==0.2.1
greenlet==2.0.0.post0
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
orjson==3.8.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.5
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.2.0
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.6
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.43
SQLAlchemy-Utils==0.38.3
starlette==0.20.4
tomli==2.0.1
typing-extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.18.3
vk-api==11.9.9
watchfiles==0.18.0
websockets==10.4
asgi-ratelimit==0.9.0